✨ feat: Split order form into a client component
